requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
peft>=0.10.0
accelerate>=0.28.0
datasets>=2.18.0
SimpleITK>=2.3.0
monai>=1.3.0
nibabel>=5.2.0
einops>=0.7.0
timm>=0.9.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
Pillow>=10.0.0
tqdm>=4.65.0
tensorboard>=2.14.0
nltk>=3.8.0
rouge-score>=0.1.2
evaluate>=0.4.0
